Refactor: views, admin doctor schedule
