+++ OUPharmacyManagementApp/mainApp/viewsets/__init__.py
@@ -2,3 +2,4 @@ from .bill import  BillViewSet
 from .doctor_schedule import DoctorScheduleViewSet
 from .time_slot import TimeSlotViewSet
 from .user import UserViewSet
+from .examination import ExaminationViewSet
